Reduce polling interval to 900 ms
The datasheet says that the sensor has to be polled every second. I was
getting CO2 measurements of 400 ppm or very close to that even when
breathing on the sensor. Reducing the polling interval to 900 ms gave me
measurements that went up much higher when breathing on the sensor. It
seems like that one second polling requirement is pretty strict, so this
change leaves more room for error.